requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
aiohttp>=3.8
Pillow>=9.0.0
psutil>=5.9.0
aioftp>=0.21.2